feat: move store sync into workflowstoresync component
